moved Util.FindByName("SuperMenuBlueprint") outside of the SpectrumMenu class constructor and created a generic method to add a menu
